Moved asyncBatchForeach, asyncForeach, and uploadFiles from gulpfile.js to tools/upload.js.
